refactor: 💡 AddMemberModal refactors
